+++ package.json
@@ -7,7 +7,8 @@
   "types": "index.d.ts",
   "scripts": {
     "dev": "vite",
-    "pack": "node scripts/publish.js",
+    "pack": "node scripts/prepublish.js pack",
+    "publish": "node scripts/prepublish.js publish",
     "build": "node scripts/build.js",
     "preview": "vite preview",
     "lint": "eslint \"**/*.{js,ts}\"",
